+++ internal/http/common/middlewares.go
@@ -9,17 +9,32 @@ import (
 )
 
 func LogRequest(h fasthttp.RequestHandler, log *zap.Logger) fasthttp.RequestHandler {
+	const headersSeparator = ": "
+
 	return func(ctx *fasthttp.RequestCtx) {
-		var (
-			startedAt = time.Now()
-			ua        = string(ctx.UserAgent())
-		)
+		var ua = string(ctx.UserAgent())
+
+		if strings.Contains(strings.ToLower(ua), "healthcheck") { // skip healthcheck requests logging
+			h(ctx)
+
+			return
+		}
+
+		var reqHeaders = make([]string, 0, 24) //nolint:gomnd
+
+		ctx.Request.Header.VisitAll(func(key, value []byte) {
+			reqHeaders = append(reqHeaders, string(key)+headersSeparator+string(value))
+		})
+
+		var startedAt = time.Now()
 
 		h(ctx)
 
-		if strings.Contains(strings.ToLower(ua), "healthcheck") { // skip healthcheck requests logging
-			return
-		}
+		var respHeaders = make([]string, 0, 16) //nolint:gomnd
+
+		ctx.Response.Header.VisitAll(func(key, value []byte) {
+			respHeaders = append(respHeaders, string(key)+headersSeparator+string(value))
+		})
 
 		log.Info("HTTP request processed",
 			zap.String("useragent", ua),
@@ -30,6 +45,8 @@ func LogRequest(h fasthttp.RequestHandler, log *zap.Logger) fasthttp.RequestHand
 			zap.String("content_type", string(ctx.Response.Header.ContentType())),
 			zap.Bool("connection_close", ctx.Response.ConnectionClose()),
 			zap.Duration("duration", time.Since(startedAt)),
+			zap.Strings("request_headers", reqHeaders),
+			zap.Strings("response_headers", respHeaders),
 		)
 	}
 }